+++ worker/src/RTC/SCTP/chunks/UnknownChunk.cpp
@@ -15,7 +15,7 @@ namespace RTC
 			MS_TRACE();
 
 			Chunk::ChunkType chunkType;
-			size_t chunkLength;
+			uint16_t chunkLength;
 			uint8_t padding;
 
 			if (!Chunk::IsChunk(buffer, bufferLength, chunkType, chunkLength, padding))
